Refactored dev-cli
