+++ src/dev-cli/logging.ts
@@ -1,0 +1,5 @@
+import { inspect } from 'node:util';
+
+export function logDeepObject(object: Record<any, any>) {
+  console.log(inspect(object, { showHidden: false, depth: null, colors: true }));
+}
