+++ src/providers/base.ts
@@ -2,6 +2,8 @@ import { Flags } from '@/entrypoint/utils/targets';
 import { Stream } from '@/providers/streams';
 import { EmbedScrapeContext, MovieScrapeContext, ShowScrapeContext } from '@/utils/context';
 
+export type MediaScraperTypes = 'show' | 'movie';
+
 export type SourcererEmbed = {
   embedId: string;
   url: string;
@@ -12,7 +14,7 @@ export type SourcererOutput = {
   stream?: Stream[];
 };
 
-export type Sourcerer = {
+export type SourcererOptions = {
   id: string;
   name: string; // displayed in the UI
   rank: number; // the higher the number, the earlier it gets put on the queue
@@ -22,15 +24,29 @@ export type Sourcerer = {
   scrapeShow?: (input: ShowScrapeContext) => Promise<SourcererOutput>;
 };
 
-export function makeSourcerer(state: Sourcerer): Sourcerer {
-  return state;
+export type Sourcerer = SourcererOptions & {
+  type: 'source';
+  disabled: boolean;
+  mediaTypes: MediaScraperTypes[];
+};
+
+export function makeSourcerer(state: SourcererOptions): Sourcerer {
+  const mediaTypes: MediaScraperTypes[] = [];
+  if (state.scrapeMovie) mediaTypes.push('movie');
+  if (state.scrapeShow) mediaTypes.push('show');
+  return {
+    ...state,
+    type: 'source',
+    disabled: state.disabled ?? false,
+    mediaTypes,
+  };
 }
 
 export type EmbedOutput = {
   stream: Stream[];
 };
 
-export type Embed = {
+export type EmbedOptions = {
   id: string;
   name: string; // displayed in the UI
   rank: number; // the higher the number, the earlier it gets put on the queue
@@ -38,6 +54,17 @@ export type Embed = {
   scrape: (input: EmbedScrapeContext) => Promise<EmbedOutput>;
 };
 
-export function makeEmbed(state: Embed): Embed {
-  return state;
+export type Embed = EmbedOptions & {
+  type: 'embed';
+  disabled: boolean;
+  mediaTypes: undefined;
+};
+
+export function makeEmbed(state: EmbedOptions): Embed {
+  return {
+    ...state,
+    type: 'embed',
+    disabled: state.disabled ?? false,
+    mediaTypes: undefined,
+  };
 }
